:sparkles: Inicio do desenvolvimento dos componentes de Toast

+++ api/src/app/controllers/ContactController.js
@@ -1,7 +1,6 @@
 const ContactsRepository = require('../repositories/ContactsRepository');
 require('express-async-errors');
-
-
+const isValidUUID = require('../utils/isValidUUID');
 class ContactController {
   async index(request, response) {
     //Index serve para LISTAR todos os registros
@@ -13,10 +12,15 @@ class ContactController {
   async show(request, response) {
     // Para obter UM registro
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id'});
+    }
+
     const contact = await ContactsRepository.findById(id);
 
     if (!contact) {
-      return response.status(404).json({ error: 'User Not Found' })
+      return response.status(404).json({ error: 'Contact Not Found' })
     }
 
     response.json(contact);
@@ -30,13 +34,22 @@ class ContactController {
       return response.status(400).json({ error: 'Name is required' })
     }
 
-    const contactByEmail = await ContactsRepository.findByEmail(email);
-    if (contactByEmail) {
-      return response.status(400).json({ error: 'This e-mail is already in use' })
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category'});
+    }
+
+    if (email){
+      const contactByEmail = await ContactsRepository.findByEmail(email);
+      if (contactByEmail) {
+        return response.status(400).json({ error: 'This e-mail is already in use' })
+      }
     }
 
     const contact = await ContactsRepository.create({
-      name, email, phone, category_id,
+      name, 
+      email: email || null, 
+      phone, 
+      category_id: category_id || null,
     });
 
     response.status(201).json(contact);
@@ -47,23 +60,36 @@ class ContactController {
     const { id } = request.params;
     const { name, email, phone, category_id } = request.body;
 
-    const contactExists = await ContactsRepository.findById(id);
-    if (!contactExists) {
-      return response.status(404).json({ error: 'User Not Found' })
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id'});
+    }
+
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category'});
     }
 
     if (!name) {
       return response.status(400).json({ error: 'Name is required' })
     }
 
-    const contactByEmail = await ContactsRepository.findByEmail(email);
-    if (contactByEmail && contactByEmail.id !== id) {
-      return response.status(400).json({ error: 'This e-mail is already in use' })
+    const contactExists = await ContactsRepository.findById(id);
+    if (!contactExists) {
+      return response.status(404).json({ error: 'User Not Found' })
+    }
+
+    if (email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        return response.status(400).json({ error: 'This e-mail is already in use' })
+      }
     }
 
     const contact = await ContactsRepository.update(id,
       {
-        name, email, phone, category_id,
+        name, 
+        email: email || null, 
+        phone, 
+        category_id: category_id || null,
       });
 
     response.json(contact);
@@ -72,6 +98,10 @@ class ContactController {
   async delete(request, response) {
     //Deletar um registro
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id'});
+    }
 
     await ContactsRepository.delete(id);
     // 204 NO CONTENT
